JAVA-1200: Upgrade LZ4 to 1.3.0.
This commit also upgrades Snappy version to 1.1.2.6 and HdrHistogram version to 2.1.9.

Besides, it also introduces new OSGi tests to validate that the following libraries can be used in an OSGi environment: LZ4, Snappy, and HdrHistogram.

As these new tests rely on system properties that might be present or not, this commit switches the Pax Exam container from "native" to "forked", because native containers are not confined enough to isolate system properties correctly.